Add functionality to save topics and opinions to csv
Functionality was added to save topics and opinions to csv files. Topics
and opinions are stored in dataframes, were the column names are the
topic indices and the index are the topic/opinion words. These
dataframes are written to disk. Todo: allow user to specify the path
where the csv files are saved.