Fix string mismatches in party names
Sometimes sting mismatches occur in party names (e.g., Groenlinks
instead of GroenLinks). The script was updated to deal with these kinds
of mismatches by doing fuzzy string matching.

Also, sometimes the party name contains a /. This results in an
additional subdirectory being created, and the data being stored in the
directory of the party name before the /. To prevent this from
happening, /'s are replaced with -'s.