Add functionality to load ranges of parameter files
The load_parameters method was changed to accept parameters to load a
single parameter file and to load ranges of parameter files, allowing
users to throw away the first part of the samples chain (burn in), or to
use a single sample as the final parameter.

Also added functionality to get the names of the parameter files. This
still needs to be integrated with the other parts of the Gibbs sampler.

Also added tests for loading the parameters and retrieving parameter
files by the new naming functionality.