Refactor for perplexity calculation
The code was refactored to prepare the perplexity calculation. It is now
possible to loop either over the train set or the test set, by setting a
parameter. This means that the Gibbs sampler can be used for both
learning topics on the train set and on the test set. This is required
for calculating perplexity on the opinion words.