Update script to calculate perplexity
The calculate perplexity script did not work in parallel for the
complete data set. Made changes to find out why it didn't work. It
appears to fail because of virtual memory usage (the script loads huge
matrices in memory). Eventually calculated results sequentially. Now
committing to make a final version of the scropt locally.