Refactor folia2cpt_input to write types of words to separate lines
The folia2cpt_input script was refactored to write different types of
words (based on pos tag) to different lines in the output text files.
This makes it easier to include different sets of words in experiments.
However, it requires to check the minimal number of words when creating
a corpus.