Add helper module for generating cpt input data
Moved the Perspective class defined in folia2cpt_corpus.py to a new
utils module that can be used to generate cpt input data. Also updated
the folia2cpt_corpus script to use this new module/class. The idea is
that the new module (class) can be reused in other scripts to generate
data.