Update experiment utils module to save results to parent dir
The results of an experiment are now saved to the parent dir of the
nTopics dir of an experiment. This was done to have one single directory
that contains all the results. This makes it easier to generate an
IPython notebook with the results, because all results can be read from
a single directory instead of two (one containing the perplexity results
and one containing the results for a specific number of topics)